types module including file-like types
